requirements.txt
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
Pygments==2.0.2
Werkzeug==0.10.4
coinbase==1.0.4
gnureadline==6.3.3
influxdb==2.8.0
jupyter==1.0.0
itsdangerous==0.24
jsonschema==2.5.1
matplotlib==1.4.3
mistune==0.7
nose==1.3.7
numpy==1.9.2
pandas==0.16.2
pyinfluxql==0.0.1
pyparsing==2.0.3
python-dateutil==2.4.2
pytz==2015.4
pyzmq==14.7.0
requests==2.7.0
scipy==0.16.0
seaborn==0.6.0
six==1.9.0
tornado==4.2.1
websockets==2.4
git+git@github.com:jjmalina/pyinfluxql.git@5f6b8558f3cf9ddf7771e2f7862b0a26e4dcaa0d#egg=pyinfluxql-dev
